fixed more imports and consolidated files within proper locations
