+++ apps/ui-sharethrift/src/components/layouts/home/pages/view-listing/components/listing-image-gallery/listing-image-gallery.container.tsx
@@ -1,9 +1,7 @@
 import { useQuery } from "@apollo/client/react";
 import { ListingImageGallery } from './listing-image-gallery.tsx';
-import {
-	ViewListingImageGalleryGetImagesDocument,
-	type ViewListingImageGalleryGetImagesQuery,
-} from '../../../../../../generated.tsx';
+import { ViewListingImageGalleryGetImagesDocument,type ViewListingImageGalleryGetImagesQuery } from '../../../../../../../generated.tsx';
+
 
 interface ListingImageGalleryContainerProps {
 	listingId: string;
