+++ apps/ui-sharethrift/src/components/layouts/home/pages/view-listing/components/listing-information/listing-information.container.graphql
